各タブをコンポーネントに分割
